feat: add diagnostic dashboard panels
