+++ .gitignore
@@ -5,6 +5,9 @@ __pycache__/
 .mypy_cache/
 .ruff_cache/
 .coverage*
+node_modules/
+playwright-report/
+test-results/
 build/
 dist/
 *.egg-info/
