+++ src/codex_usage_tracker/dashboard.py
@@ -76,6 +76,16 @@ DASHBOARD_SCRIPT_ASSETS = (
     ("actions_script_src", "__ACTIONS_SCRIPT_SRC__", "dashboard_actions.js"),
     ("live_script_src", "__LIVE_SCRIPT_SRC__", "dashboard_live.js"),
     ("events_script_src", "__EVENTS_SCRIPT_SRC__", "dashboard_events.js"),
+    (
+        "diagnostics_snapshots_script_src",
+        "__DIAGNOSTICS_SNAPSHOTS_SCRIPT_SRC__",
+        "dashboard_diagnostics_snapshots.js",
+    ),
+    (
+        "diagnostics_facts_script_src",
+        "__DIAGNOSTICS_FACTS_SCRIPT_SRC__",
+        "dashboard_diagnostics_facts.js",
+    ),
     ("diagnostics_script_src", "__DIAGNOSTICS_SCRIPT_SRC__", "dashboard_diagnostics.js"),
     (
         "call_diagnostics_script_src",
@@ -320,6 +330,8 @@ def render_dashboard_html(
     actions_script_src: str | None = None,
     live_script_src: str | None = None,
     events_script_src: str | None = None,
+    diagnostics_snapshots_script_src: str | None = None,
+    diagnostics_facts_script_src: str | None = None,
     diagnostics_script_src: str | None = None,
     call_diagnostics_script_src: str | None = None,
     call_investigator_script_src: str | None = None,
@@ -347,6 +359,8 @@ def render_dashboard_html(
         "actions_script_src": actions_script_src,
         "live_script_src": live_script_src,
         "events_script_src": events_script_src,
+        "diagnostics_snapshots_script_src": diagnostics_snapshots_script_src,
+        "diagnostics_facts_script_src": diagnostics_facts_script_src,
         "diagnostics_script_src": diagnostics_script_src,
         "call_diagnostics_script_src": call_diagnostics_script_src,
         "call_investigator_script_src": call_investigator_script_src,
