+++ src/codex_usage_tracker/plugin_data/dashboard/dashboard_responsive.css
@@ -30,6 +30,9 @@
       .pager { width: auto; }
       .action-status { min-width: 0; }
       .preset-card { grid-template-columns: 1fr; }
+      .diagnostics-toolbar { align-items: stretch; flex-direction: column; }
+      .diagnostics-toolbar .toolbar-button { width: 100%; }
+      .diagnostics-snapshot-grid { grid-template-columns: 1fr; }
       .diagnostics-section-header { display: grid; }
       .diagnostics-readout .diagnostics-note { grid-column: auto; }
       body[data-active-view="call"] .call-investigator {
