+++ src/codex_usage_tracker/plugin_data/dashboard/dashboard_template.html
@@ -171,6 +171,8 @@
   <script src="__ACTIONS_SCRIPT_SRC__"></script>
   <script src="__LIVE_SCRIPT_SRC__"></script>
   <script src="__EVENTS_SCRIPT_SRC__"></script>
+  <script src="__DIAGNOSTICS_SNAPSHOTS_SCRIPT_SRC__"></script>
+  <script src="__DIAGNOSTICS_FACTS_SCRIPT_SRC__"></script>
   <script src="__DIAGNOSTICS_SCRIPT_SRC__"></script>
   <script src="__CALL_DIAGNOSTICS_SCRIPT_SRC__"></script>
   <script src="__CALL_INVESTIGATOR_SCRIPT_SRC__"></script>
